rename userprofile to entry-form and move the checkbox components into the component folder
