+++ src/app/edit-profile/page.tsx
@@ -4,8 +4,8 @@ import React, { ChangeEvent, useEffect, useState } from 'react'
 import axios from 'axios';
 import styles from '../userinfo/UserInfo.module.css'
 import { UserAuth } from '../context/AuthContext';
-import Checkbox from '../utils/Checkbox';
-import LearningCheckbox from '../utils/Learning-Checkbox';
+import Checkbox from '../components/elements/Checkbox';
+import LearningCheckbox from '../components/elements/Learning-Checkbox';
 import langCall from '../utils/langCheckFunc';
 import { useRouter } from 'next/navigation';
